requirements.txt
Django==3.0.7
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-modeltranslation==0.15
Pillow==7.1.2